Add files via upload
An Insurance firm providing tour insurance is facing higher claim frequency. The management decides to collect data from the past few years. We were assigned the task to make a model which predicts the claim status and provide recommendations to management. Used CART, RF & ANN and compare the models' performances in train and test sets.
Data Dictionary for Market Segmentation:
1. Target: Claim Status (Claimed)
2. Code of tour firm (Agency_Code)
3. Type of tour insurance firms (Type)
4. Distribution channel of tour insurance agencies (Channel)
5. Name of the tour insurance products (Product)
6. Duration of the tour (Duration)
7. Destination of the tour (Destination)
8. Amount of sales of tour insurance policies (Sales)
9. The commission received for tour insurance firm (Commission)
10. Age of insured (Age)